[FIRRTL] Move FieldRef utils into FIRRTLUtils.h, NFC
These helpers predate the existance of the FIRRTLUtils.h header file and
were just hanging out in FIRRTLDialect.h.
